refactor: use proper tuple type instead of unknown[] in and/or implementation
