+++ src/comparison.ts
@@ -4,7 +4,17 @@ import { type Argument, Operation } from './operation';
  * Operator function reference type.
  * Accepts one or more arguments and returns an {@link Operation}.
  */
-type Operator = ((value: Argument) => Operation) | ((...args: Argument[]) => Operation);
+export type Operator = ((value: Argument) => Operation) | ((...args: Argument[]) => Operation);
+
+/**
+ * A tuple representing a field comparison without creating a {@link Comparison} instance.
+ * Useful as a shorthand inside {@link and} and {@link or}.
+ *
+ * @example
+ * ['field1', eq, 'val']
+ * ['field2', inList, 'foo', 'bar', 'baz']
+ */
+export type ComparisonTuple = [fieldName: string, operator: Operator, ...values: Argument[]];
 
 /**
  * Comparison groups delimiters
@@ -56,6 +66,8 @@ export function comparison(selector: string, operation: Operation): Comparison;
  * // Multiple arguments:
  * import {inList} from 'rsql-builder';
  * const comp2 = comparison('field2', inList, 'foo', 'bar');  // 'field2=in=(foo,bar)'
+ *
+ * See also {@link ComparisonTuple} for use inside {@link and} / {@link or}.
  */
 export function comparison(selector: string, operator: Operator, ...values: Argument[]): Comparison;
