+++ tests/or.spec.js
@@ -1,6 +1,6 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert/strict';
-import { or, eq, inList, comparison } from '../dist/index.js';
+import { or, cmp, eq, inList, comparison } from '../dist/index.js';
 
 describe('or()', () => {
   it('should return or-expression string', () => {
@@ -21,6 +21,16 @@ describe('or()', () => {
     it('should mix tuples, strings and comparisons', () => {
       const query = or(['genres', inList, 'sci-fi', 'action'], comparison('director', eq('Nolan')), 'year>=2000');
       assert.strictEqual(query, 'genres=in=(sci-fi,action),director==Nolan,year>=2000');
+    });
+
+    it('should compose with or()', () => {
+      const query = or(comparison('genres', inList, 'sci-fi', 'action'), cmp('director', eq('Nolan')));
+      assert.strictEqual(query, 'genres=in=(sci-fi,action),director==Nolan');
+    });
+
+    it('should compose with or() using tuple syntax', () => {
+      const query = or(['field1', eq, 'val'], ['field2', inList, 'foo', 'bar']);
+      assert.strictEqual(query, 'field1==val,field2=in=(foo,bar)');
     });
   });
 });
